Add function to connect to an arbitrary libvirt URI
When opening a connection with the libvirt client, the provided URI is
first used to set up the socket used to communicate with libvirtd. The
actual URI passed to libvirtd (i.e. when calling the ConnectToURI()
method) contains only a subset of the original, since relative to
libvirtd the underlying virt provider is always local.

Create a new ConnectToURI() function that returns a connected Libvirt
using the appropriate dialer and remote URI.